requirements.txt
sounddevice
numpy
scipy
noisereduce
webrtcvad-wheels
faster-whisper
deep-translator
edge-tts